if cuda is not available, use cpu

+++ src/check_nii_dims.sh
@@ -17,7 +17,7 @@ echo "Current dimensions are $dim1 $dim2 $dim3"
 
 # if dim1 is an odd number, then add a slice with zeros to dim1
 if ((dim1 % 2 != 0)); then
-    echo "adding slice with zeros to dim1 of $2"
+    echo "adding slice with zeros to dim1 of $1"
     # extract a slice from the input.nii.gz file
     fslroi $1 zero_slice.nii.gz 0 1 0 $dim2 0 $dim3
     fslmaths zero_slice.nii.gz -mul 0 zero_slice.nii.gz
@@ -26,7 +26,7 @@ fi
 
 # if dim2 is an odd number, then add a slice with zeros to dim2
 if ((dim2 % 2 != 0)); then
-    echo "adding slice with zeros to dim2 of $2"
+    echo "adding slice with zeros to dim2 of $1"
     # extract a slice from the input.nii.gz file
     fslroi $1 zero_slice.nii.gz 0 $dim1 0 1 0 $dim3
     fslmaths zero_slice.nii.gz -mul 0 zero_slice.nii.gz
@@ -35,7 +35,7 @@ fi
 
 # if dim3 is an odd number, then add a slice with zeros to dim3
 if ((dim3 % 2 != 0)); then
-    echo "adding slice with zeros to dim3 of $2"
+    echo "adding slice with zeros to dim3 of $1"
     # extract a slice from the input.nii.gz file
     fslroi $1 zero_slice.nii.gz 0 $dim1 0 $dim2 0 1
     fslmaths zero_slice.nii.gz -mul 0 zero_slice.nii.gz
@@ -48,4 +48,8 @@ dim2_new=$(fslhd $2 | grep -w dim2 | awk '{print $2}')
 dim3_new=$(fslhd $2 | grep -w dim3 | awk '{print $2}')
 
 echo "Finished: new dimensions are $dim1_new $dim2_new $dim3_new"
-rm zero_slice.nii.gz
+
+# remove if exists
+if [ -f zero_slice.nii.gz ]; then
+    rm zero_slice.nii.gz
+fi
